+++ src/styles/_inputs.scss
@@ -5,6 +5,7 @@
     border: 1px solid colors.$text-gray-3;
     border-radius: 8px;
     display: flex;
+    padding: 0 8px;
     align-items: center;
 
     &:hover {
@@ -29,7 +30,6 @@
       height: 40px;
       border: 0;
       width: 100%;
-      padding: 0 8px;
       border-radius: 8px;
 
       &:focus {
@@ -75,9 +75,12 @@
       }
     }
 
-    &.seam-search-text-field {
+    &.seam-large {
+      padding: 0 12px;
+
       .seam-text-field-input {
-        width: 100%;
+        height: 48px;
+        font-size: 16px;
       }
     }
   }
